backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets==13.1
google-adk==1.0.0
google-genai>=1.14.0
google-cloud-firestore==2.19.0
python-dotenv==1.0.1
aiohttp==3.10.10
pydantic>=2.9.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-timeout==2.3.1
httpx>=0.28.1
